Refactor server into routes and controllers
