+++ src/server/routes/userRoutes.js
@@ -38,78 +38,16 @@ router.post('/login',
 // Returns 200 on success
 router.get('/logout', auth.requireLoggedIn, userController.logoutUser);
 
-
-
-
-
-
-
-
-
-
-
-
-
 //Delete a user
-router.route('/ban/:id').delete((req,res) => {
-    UserModel.findByIdAndDelete(req.params.id)
-        .then(() => {
-            console.log("User Deleted in the database");
-            res.end();
-        })
-        .catch(error => console.log("Error"));
-})
+router.delete('/ban/:id', userController.deleteUser);
 
 //Update user's information
-router.route('/update/:email').post((req,res) => {
-    var name;
-    var email;
-    console.log(req.params.email)
-    UserModel.findOneAndUpdate({email: req.params.email})
-        .then(user => {
-            user.email = req.body.email;
-            email = user.email;
-            user.name = req.body.name;
-            name = user.name;
-            user.save()
-                .then(() => {
-                    console.log("User was updated in the database");
-                    res = user;
-                })
-                
-                .catch(error => console.log(error));
-        } )
-        .catch(error => res.json("Error finding user"));   
-});
-
-// ProductModel.findById(req.params.id)
-// .then(product => {
-
-//     const index = product.inUserCart.indexOf(req.body.email);
-//     if (index > -1) {
-//         product.inUserCart.splice(index, 1); // 2nd parameter means remove one item only
-//     }
-//     product.save()
-//         .then(() => console.log("Product was updated in the database"))
-//         .catch(error => console.log(error));
-// } )
+router.post('/update/:email', userController.updateUser);
 
 //Update user's password
-router.route('/updatePassword/:id').post((req,res) => {
-    UserModel.findOneAndUpdate({email: req.params.email})
-        .then(user => {
-            
-            const saltAndHash = auth.generateNewPassword(req.body.hashedPassword);
-            user.hashedPassword = saltAndHash.hash
-            user.salt = saltAndHash.salt
-            user.save()
-                .then(() => {
-                    console.log("User password was updated in the database");
-                    res.end();
-                })
-                .catch(error => console.log(error)); 
-        } )
-        .catch(error => res.json("Error finding user"))
-});
+router.post('/updatePassword/:id', userController.updatePassword);
+
+//Api endpoint for users
+router.get('/search', userController.getCurrentUser);
 
 module.exports = router;
